web_project/requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
mysqlclient==2.0.3
requests==2.26.0

